requirements.txt
Flask>=3.0.0
feedparser>=6.0.10
requests>=2.31.0
python-dateutil>=2.8.2
readability-lxml>=0.8.1
beautifulsoup4>=4.12.3
lxml>=5.2.1
cssselect>=1.2.0
openai>=1.40.0
python-dotenv>=1.0.1
gunicorn>=21.2.0